requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Deprecated==1.2.13
frozenlist==1.3.1
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gspread==5.4.0
httplib2==0.20.4
idna==3.3
infi.systray==0.1.12
keyboard==0.13.5
MouseInfo==0.1.3
multidict==6.0.2
oauth2client==4.1.3
oauthlib==3.2.0
Pillow==9.2.0
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyDrive==1.3.1
PyGetWindow==0.0.9
PyGithub==1.55
PyJWT==2.4.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pystray==0.19.4
pytweening==1.0.4
pywin32==304
PyYAML==6.0
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
yarl==1.8.1
